tuls-hub-api/requirements.txt
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.2.25
chardet==7.4.0.post2
charset-normalizer==3.4.7
click==8.3.2
cssselect==1.4.0
Flask==3.1.3
flask-cors==6.0.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
lxml_html_clean==0.4.4
MarkupSafe==3.0.3
python-docx==1.2.0
readability-lxml==0.8.4.1
requests==2.33.1
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
